Introduced autoForward support for userInfo
Added a variation of `performSegueWithIdentifier` that supports an
`autoForward`. At this time, the intent is to allow the segue target to
be a `UINavigationController`, but the `userInfo` is passed instead to
the root view controller.

If other classes need similar support, the logic could be different
inside the swizzled `prepareForSegue`, but the concept would still be
forwarding `userInfo` to a child controller.